requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.5
langchain-community==0.2.5
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.3
sqlmodel==0.0.18
pandas==2.2.2
sqlalchemy==2.0.30
streamlit==1.36.0
tiktoken==0.7.0
ollama==0.2.1
sentence-transformers==2.7.0
chromadb==0.5.4
langchain-groq